Coverity Fixes: Several fixes to coverity issues in some sysbios-rpmsg files.
The coverity fixes address (or supress) the errors reported by coverity.

NOTE: the virtio_ring.h file copyright header triggers a tisb filter check;
but this is copyrighted by Rusty Russel, so commit was entered with the check
disabled.

+++ packages/ti/trace/SysMin.c
@@ -188,7 +188,7 @@ Void SysMin_putch(Char ch)
 #else
         if (module->getTime == TRUE) {
 #endif
-            uSec  = Clock_getTicks() * Clock_tickPeriod;
+            uSec  = Clock_getTicks() * (UInt64)Clock_tickPeriod;
             SysMin_output('[');
             if (uSec) {
                 sprintf(timeStamp, "%17llu\0", uSec);
